Fixed the circular importing problem in last commit by introducing seppy/tools/util, that holds helper functions needed by the Event class.
