+++ seppy/loader/soho.py
@@ -15,7 +15,7 @@ from sunpy.net import Fido
 from sunpy.net import attrs as a
 from sunpy.timeseries import TimeSeries
 
-from seppy.tools import resample_df
+from seppy.tools.util import resample_df
 
 
 def _get_metadata(dataset, path_to_cdf):
