+++ seppy/loader/wind.py
@@ -14,7 +14,7 @@ import warnings
 from sunpy.net import Fido
 from sunpy.net import attrs as a
 
-from seppy.tools import resample_df
+from seppy.tools.util import resample_df
 
 def _date2str(date):
     year = str(date)[0:4]
